Seperate out logger into different files
